demo_project/requirements.txt
snowflake-connector-python
streamlit
plotly
mage-ai
itsdangerous